Convert course-helpers to WebSOC class
