+++ apps/antalmanac/src/components/RightPane/CoursePane/CoursePaneRoot.tsx
@@ -6,7 +6,7 @@ import CourseRenderPane from './CourseRenderPane';
 import SearchForm from './SearchForm/SearchForm';
 import analyticsEnum, { logAnalytics } from '$lib/analytics';
 import { openSnackbar } from '$actions/AppStoreActions';
-import { clearCache } from '$lib/course-helpers';
+import WebSOC from '$lib/websoc';
 
 function RightPane() {
     const [key, forceUpdate] = useReducer((currentCount) => currentCount + 1, 0);
@@ -33,7 +33,7 @@ function RightPane() {
             category: analyticsEnum.classSearch.title,
             action: analyticsEnum.classSearch.actions.REFRESH,
         });
-        clearCache();
+        WebSOC.clearCache();
         forceUpdate();
     }, []);
